QSDK:PROTOCOL: hostapd: Fix debug compilation issue
Currently, the following compilation issue is seen in below line:
format '%ld' expects argument of type 'long int', but argument 3 has type 'int'
"MLD: Unexpected MLE length: %ld != %zu",  mle_pos - pos, mle_len);

Fix this by adding the appropriate type cast for the (mle_pos - pos).

Patch-dependency: None
TAG: GENERAL

+++ src/ap/ieee802_11_eht.c
@@ -2131,7 +2131,7 @@ hostapd_send_link_reconf_resp(struct hostapd_data *hapd,
 		if ((size_t) (mle_pos - pos) != mle_len) {
 			wpa_printf(MSG_DEBUG,
 				   "MLD: Unexpected MLE length: %ld != %zu",
-				   mle_pos - pos, mle_len);
+				   (long) (mle_pos - pos), mle_len);
 			reject_all = true;
 			goto reject_all_req;
 		}
